(create-route-controllers): create endpoints controllers
- create route controllers for all routes

[Finishes #158662816]
