--- .nyc_output/8ce9c26f4b20f06107ce8b463e35705f.json
@@ -1,1 +1,0 @@
-{}
