+++ server/controllers/index.js
@@ -1,0 +1,14 @@
+import getRides from './rides/getRides';
+import getSingleRide from './rides/getSingleRide';
+import createRide from './rides/createRide';
+import createRideRequest from './rides/joinRide';
+import getRideRequests from './rides/getRideRequest';
+
+
+export default {
+  getRides,
+  getSingleRide,
+  createRide,
+  createRideRequest,
+  getRideRequests,
+};
