+++ server/controllers/rides/getRides.js
@@ -1,0 +1,7 @@
+import db from '../../data/rides';
+
+const getRides = (req, res) => {
+  res.json(db);
+};
+
+export default getRides;
